use uint64 as note ID

+++ server/internal/core/domain/round_repo.go
@@ -35,7 +35,7 @@ type VtxoRepository interface {
 }
 
 type NoteRepository interface {
-	Contains(ctx context.Context, id uint32) (bool, error)
-	Add(ctx context.Context, id uint32) error
+	Contains(ctx context.Context, id uint64) (bool, error)
+	Add(ctx context.Context, id uint64) error
 	Close()
 }
